+++ src/main/java/dev/lambdaurora/lambdabettergrass/mixin/MainMixin.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright © 2023 LambdAurora <email@lambdaurora.dev>
+ *
+ * This file is part of LambdaBetterGrass.
+ *
+ * Licensed under the MIT license. For more information,
+ * see the LICENSE file.
+ */
+
+package dev.lambdaurora.lambdabettergrass.mixin;
+
+import net.minecraft.client.main.Main;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(Main.class)
+public final class MainMixin {
+	@Inject(method = "*([Ljava/lang/String;Z)V", at = @At("HEAD"), cancellable = true, remap = false)
+	private static void lbg$onMain(String[] args, boolean bl, CallbackInfo ci) {
+		ci.cancel();
+	}
+}